Updating the game drawables
- Updating the game assets
- Updating the icon
- Added Power Item class
