+++ android/AndroidManifest.xml
@@ -9,7 +9,7 @@ xmlns:tools="http://schemas.android.com/tools">
     <application
         android:allowBackup="true"
         android:fullBackupContent="true"
-        android:icon="@drawable/ic_launcher"
+        android:icon="@mipmap/ic_launcher"
         android:isGame="true"
         android:label="@string/app_name"
         android:theme="@style/AppTheme"
