+++ core/src/com/hesham0_0/spaceship/PowerType.java
@@ -1,0 +1,7 @@
+package com.hesham0_0.spaceship;
+
+public enum PowerType {
+    RINGS,
+    BULLETS,
+
+}
