+++ core/src/com/hesham0_0/spaceship/SpaceshipContactCallback.java
@@ -2,16 +2,16 @@ package com.hesham0_0.spaceship;
 
 import com.badlogic.gdx.math.Vector2;
 import com.hesham0_0.spaceship.models.Wall;
-import com.hesham0_0.spaceship.models.spaceshipBullet;
-import com.hesham0_0.spaceship.models.spaceshipRing;
-import com.hesham0_0.spaceship.models.spaceshipRock;
+import com.hesham0_0.spaceship.models.SpaceshipBullet;
+import com.hesham0_0.spaceship.models.SpaceshipRing;
+import com.hesham0_0.spaceship.models.SpaceshipRock;
 
 public interface SpaceshipContactCallback {
-    void bulletRockCollision(spaceshipRock rock, spaceshipBullet bullet, Vector2 contactPosition);
+    void bulletRockCollision(SpaceshipRock rock, SpaceshipBullet bullet, Vector2 contactPosition);
 
-    void ringRocksCollision(spaceshipRing ring, spaceshipRock rock, Vector2 contactPosition);
+    void ringRocksCollision(SpaceshipRing ring, SpaceshipRock rock, Vector2 contactPosition);
 
-    void wallRocksCollision(Wall wall, spaceshipRock rock);
+    void wallRocksCollision(Wall wall, SpaceshipRock rock);
 
-    void wallBulletCollision(Wall wall, spaceshipBullet bullet);
+    void wallBulletCollision(Wall wall, SpaceshipBullet bullet);
 }
